FDP-1879 Fix for publication venue
This is a fix for publication venue not being returned in vivo
widgets. Widgets was originally just pulling the venue relationship, but
not all journals had a relationship (which means there weren't
corresponding journal entries in Elements). This changes looks for the
journal title right on the publication and not linked.