FDP-1844: Add Artistic Works
Events currently exist in the sparql with all features, but only
returning a comma-separated list of uris. We need to decide how much
inforamtion we want to provide for these in the API. If we are not
passing it through, we should get rid of the SPARQL that gets details on
them.

Account for work date precisions