FDP-2032: Fix grant date filtering logic
A grant with grant dates grant_start - grant_end is within time_start -
time_end if the time_start is before or equal to grant_end and time_end
is after or equal to grant_start. This is also correct for grants
without a specified start or end, so long as that means it goes for
infinity in the unspecified direction.