FDP-2099 Cleaned up logging and exceptions
The reindex futures were sporadically stopping (or hanging) on
test. Maybe an exception was getting eaten as part of the PersonIndexer
so I added another generic case statement in the catch.

Fixed logback so it not logging duplicate entries into the log. Also,
the log should now be rotated in the logging directory.

General logging clean up. Added additional debugging logging.

Turned up logging to debug for PersonIndexer and VivoSolrIndexer.