FDP-2173: Add mentorship overview
This could be put into the builder as a separate item, like the
overview. Or, we could try to construct something in the builder that
mashes together the mentorship overview and the availabilities, analogously to
what we do in Vivo itself. But for now, this is all that is really
needed.