1.7: Add an authorshipType
to distinguish between authorships, translatorships, and editorships

Note that I tried to use strafter function in the sparql query (to
return only 'Authorship', instead of the fully qualified
vivo:Authorship, for example), but our version of Jena does not seem to
allow it. I tried to update Jena, but ran into a rabbit hole of
difficulties, possibly because SDB is no longer well-supported.