Comment out tests that cannot auto-run
There are two issues that prevent these otherwise useful tests from
passing on the CI server.

The first is that they require an external solr server AND that they
have to run once, the solr server has to be restarted, and then they
have to run again in order to obtain a valid result. (I use
bin/solr_dev.sh {start|stop} to start and stop the solr server.)

The second is that they each try to setup an in-memory h2 database. They
follow the pattern set in the VivoIndexerSpec. This, however, only works
the first time in a test run. We need a better test database setup and
cleanup system.