FDP-1470: Prevent over indexing based on type
This is a Duke specific hack and will be refactored. For now, use our
uri naming conventions to pick out orgs and people and re-index them
without checking the rest of widgets for related enities to also
reindex. Ulitmately we must make the type determination external to
widgets and potentially expose different endpoints for each case:
people, organizations, and all others.