FDP-3471: Update endpoint fields
Add vivoType and updatedAt to top-level response.

Remove 'contact' section. It is identical to 'overview'. Both return the
main attributes of the person. Overview is a better name for this.

Minor refactor: Get the context once and pass that instance to each call
to determine an endpoint url

Sort sections in source code for readability